requirements.txt
uvicorn
numpy 
Flask 
urllib3 
Flask-Cors 
PyYAML 
fastapi
GitPython 
PyGithub
chromadb
openai
tiktoken
tenacity
python-dotenv
esprima